### What's in this PR?
-  Added a landing page (**Landing.tsx**) rendered before the lab UI.
- Used only the repo’s tactical classes/tokens (bg-tactical-*, text-tactical-*, btn-tactical-*, tactical-card, etc.) for consistent styling.
- No new dependencies.
### Work Process
**1. Initial Implementation**
- Created hero, features, CTA sections.
- Verified console is clean, responsive layout works across breakpoints, and accessibility basics ( focusable buttons, visible focus states).

**2.** **Performance inspection (Lighthouse, mobile throttling)**

- Initial metrics showed very slow paint due to editor/Monaco loading eagerly with the landing.

**3.**  **Optimization step**

- Converted Editor and MarkdownRenderer to React.lazy + Suspense
- These bundles now load only after clicking “Start.”
- Prevents Monaco/markdown from blocking the first paint on the landing.
### Results
### Lighthouse (mobile, throttled)

**Before optimization (initial landing):**
- FCP ≈ 15s
- LCP ≈ 28s
- Speed Index ≈ 18s

**After optimization (lazy-loaded editor/markdown):**
- FCP ≈ 2.6s
- LCP ≈ 2.6s
- Speed Index ≈ 3.1s
### Quality Checks
- ✅ Console clean.

-  ✅ Responsive at 360 / 768 / 1024 / 1440 px.

- ✅ No new dependencies.

- ✅ Uses design system tokens.

- ✅ Production build tested (pnpm run build && pnpm run preview).

<img width="2048" height="1854" alt="localhost_4300_" src="https://github.com/user-attachments/assets/52b67186-9439-44cd-854f-a10649e73f3b" />